+++ app/src/adapters/electronTypes.ts
@@ -13,6 +13,7 @@ export type {
   Event,
   HandlerDetails,
   IpcMainEvent,
+  IpcMainInvokeEvent,
   IpcRenderer,
   Menu,
   MenuItem,