+++ src/options/normalizeUrl.ts
@@ -1,35 +1,14 @@
-import * as url from 'url';
-
 import * as log from 'loglevel';
 
-function hasProtocol(inputUrl: string): boolean {
-  return /^[a-zA-Z][a-zA-Z\d+\-.]*:/.test(inputUrl);
-}
-
-function appendProtocol(inputUrl: string): string {
-  if (!hasProtocol(inputUrl)) {
-    const urlWithProtocol = `https://${inputUrl}`;
-    log.warn(
-      `URL "${inputUrl}" lacks a protocol.`,
-      `Will try to parse it as HTTPS: "${urlWithProtocol}".`,
-      `Please pass "http://${inputUrl}" if this is what you meant.`,
-    );
-    return urlWithProtocol;
-  }
-  return inputUrl;
-}
+import { normalizeUrl as normalizeUrlShared } from '../../shared/lib/src/options/normalizeUrl';
 
 export function normalizeUrl(urlToNormalize: string): string {
-  const urlWithProtocol = appendProtocol(urlToNormalize);
-
-  let parsedUrl: url.URL;
   try {
-    parsedUrl = new url.URL(urlWithProtocol);
+    const normalizedUrl = normalizeUrlShared(urlToNormalize);
+    log.debug(`Normalized URL ${urlToNormalize} to:`, normalizedUrl);
+    return normalizedUrl;
   } catch (err: unknown) {
     log.error('normalizeUrl ERROR', err);
-    throw new Error(`Your url "${urlWithProtocol}" is invalid`, { cause: err });
+    throw err;
   }
-  const normalizedUrl = parsedUrl.toString();
-  log.debug(`Normalized URL ${urlToNormalize} to:`, normalizedUrl);
-  return normalizedUrl;
 }